requirements.txt
discord.py==1.6.0
requests==2.26.0
python-dotenv==0.19.2
